Move DownloadInfoManager to rocket package
