+++ app/src/main/java/org/mozilla/rocket/download/DownloadInfoRepository.kt
@@ -1,9 +1,9 @@
 package org.mozilla.rocket.download
 
 import android.app.DownloadManager
-import org.mozilla.rocket.download.data.DownloadInfo
-import org.mozilla.focus.download.DownloadInfoManager
 import org.mozilla.focus.telemetry.TelemetryWrapper
+import org.mozilla.rocket.download.data.DownloadInfo
+import org.mozilla.rocket.download.data.DownloadInfoManager
 import org.mozilla.threadutils.ThreadUtils
 
 class DownloadInfoRepository {
